jest folder structure and pass tests
